+++ Modules/Microsoft365DSC/DSCResources/MSFT_IntuneDeviceControlPolicyWindows10/MSFT_IntuneDeviceControlPolicyWindows10.psm1
@@ -137,6 +137,16 @@ function Get-TargetResource
         $AllowFullScanRemovableDriveScanning,
 
         [Parameter()]
+        [ValidateSet('1', '2')]
+        [System.String]
+        $DefaultEnforcement,
+
+        [Parameter()]
+        [ValidateSet('0', '1')]
+        [System.String]
+        $DeviceControlEnabled,
+
+        [Parameter()]
         [ValidateSet('0', '1')]
         [System.String]
         $AllowDirectMemoryAccess,
@@ -515,6 +525,16 @@ function Set-TargetResource
         $AllowFullScanRemovableDriveScanning,
 
         [Parameter()]
+        [ValidateSet('1', '2')]
+        [System.String]
+        $DefaultEnforcement,
+
+        [Parameter()]
+        [ValidateSet('0', '1')]
+        [System.String]
+        $DeviceControlEnabled,
+
+        [Parameter()]
         [ValidateSet('0', '1')]
         [System.String]
         $AllowDirectMemoryAccess,
@@ -834,6 +854,16 @@ function Test-TargetResource
         [ValidateSet('0', '1')]
         [System.String]
         $AllowFullScanRemovableDriveScanning,
+
+        [Parameter()]
+        [ValidateSet('1', '2')]
+        [System.String]
+        $DefaultEnforcement,
+
+        [Parameter()]
+        [ValidateSet('0', '1')]
+        [System.String]
+        $DeviceControlEnabled,
 
         [Parameter()]
         [ValidateSet('0', '1')]